Fix all broken tests
