+++ src/prefixer/ModuleManager.spec.ts
@@ -2,7 +2,7 @@ import { ModuleManager } from "./ModuleManager";
 import { expect } from "chai";
 import * as path from 'path';
 import { util } from "../util";
-import { file, fsEqual, createProjects, DepGraphNode } from "../TestHelpers.spec";
+import { file, fsEqual, createProjects, DepGraphNode, trim } from "../TestHelpers.spec";
 
 const hostDir = path.join(process.cwd(), '.tmp', 'hostApp');
 
@@ -29,6 +29,7 @@ describe('ModuleManager', function () {
         await Promise.all(
             manager.modules.map(x => x.init())
         );
+        return manager.modules;
     }
 
     describe('getReducedDependencies', () => {
@@ -163,32 +164,149 @@ describe('ModuleManager', function () {
         });
     });
 
-    /**
-       * This test converts the dependency name "module1" to "module2", and names this package "module1"
-       */
-    it('handles module prefix swapping', async () => {
-        createDependencies([{
-            alias: 'logger',
-            name: 'simple-logger',
-            dependencies: [{
+    describe('process', () => {
+
+        /**
+         * This test converts the dependency name "module1" to "module2", and names this package "module1"
+         */
+        it('handles module prefix swapping', async () => {
+            createDependencies([{
                 alias: 'logger',
-                name: 'complex-logger'
-            }],
-        }]);
+                name: 'simple-logger',
+                dependencies: [{
+                    alias: 'logger',
+                    name: 'complex-logger'
+                }],
+            }]);
 
-        //simple-logger calls method from complex-logger, aliased as `logger`
-        file(`${hostDir}/node_modules/logger/source/main.brs`, `
-            sub WriteToLog(message)
-                return logger_writeToLog(message)
-            end sub
-        `);
-        await process();
+            //simple-logger calls method from complex-logger, aliased as `logger`
+            file(`${hostDir}/node_modules/logger/source/main.brs`, `
+                sub WriteToLog(message)
+                    return logger_writeToLog(message)
+                end sub
+            `);
+            await process();
 
-        fsEqual(`${hostDir}/source/roku_modules/logger/main.brs`, `
-            sub logger_WriteToLog(message)
-                return complexlogger_v1_writeToLog(message)
-            end sub
-        `);
+            fsEqual(`${hostDir}/source/roku_modules/logger/main.brs`, `
+                sub logger_WriteToLog(message)
+                    return complexlogger_v1_writeToLog(message)
+                end sub
+            `);
+        });
+
+        it('applies a prefix to all functions of a program', async () => {
+            let [logger] = await createDependencies([{
+                name: 'logger'
+            }]);
+
+            file(`${logger.rootDir}/source/main.brs`, `
+                sub main()
+                    SanitizeText("123")
+                end sub
+                sub SanitizeText(text as string)
+                    PrintMessage("Sanitizing text: " + text)
+                end sub
+            `);
+            file(`${logger.rootDir}/source/lib.brs`, `
+                sub PrintMessage(message)
+                    print message
+                end sub
+            `);
+            await process();
+
+            fsEqual(`${hostDir}/source/roku_modules/logger/main.brs`, `
+                sub main()
+                    logger_SanitizeText("123")
+                end sub
+                sub logger_SanitizeText(text as string)
+                    logger_PrintMessage("Sanitizing text: " + text)
+                end sub
+            `);
+
+            fsEqual(`${hostDir}/source/roku_modules/logger/lib.brs`, `
+                sub logger_PrintMessage(message)
+                    print message
+                end sub
+            `);
+        });
+
+        it('applies a prefix to components and their usage', async () => {
+            let [logger] = await createDependencies([{
+                name: 'logger'
+            }]);
+
+            file(`${logger.rootDir}/components/Component1.xml`, trim`
+                <?xml version="1.0" encoding="utf-8" ?>
+                <component name="Component1" extends="Component2" >
+                </component>
+            `);
+            file(`${logger.rootDir}/components/Component2.xml`, trim`
+                <?xml version="1.0" encoding="utf-8" ?>
+                <component name="Component2" extends="Task" >
+                    <children>
+                        <Component1 />
+                        <Component2>
+                        </Component2>
+                    </children>
+                </component>
+            `);
+            file(`${logger.rootDir}/source/main.brs`, `
+                sub main()
+                    comp = CreateObject("rosgnode", "Component1")
+                    comp.CreateChild("Component2")
+                end sub
+            `);
+            await process();
+
+            fsEqual(`${hostDir}/components/roku_modules/logger/Component1.xml`, `
+                <?xml version="1.0" encoding="utf-8" ?>
+                <component name="logger_Component1" extends="logger_Component2" >
+                </component>
+            `);
+            fsEqual(`${hostDir}/components/roku_modules/logger/Component2.xml`, `
+                <?xml version="1.0" encoding="utf-8" ?>
+                <component name="logger_Component2" extends="Task" >
+                    <children>
+                        <logger_Component1 />
+                        <logger_Component2>
+                        </logger_Component2>
+                    </children>
+                </component>
+            `);
+
+            fsEqual(`${hostDir}/source/roku_modules/logger/main.brs`, `
+                sub main()
+                    comp = CreateObject("rosgnode", "logger_Component1")
+                    comp.CreateChild("logger_Component2")
+                end sub
+            `);
+        });
+
+        it('renames dependency prefixes', async () => {
+            let [logger] = await createDependencies([{
+                alias: 'logger',
+                name: '@alpha/logger',
+                dependencies: [{
+                    alias: 'logger',
+                    name: '@bravo/printer'
+                }]
+            }]);
+
+            file(`${logger.rootDir}/source/main.brs`, `
+                sub PrintValue(value)
+                    print logger_writeLine(value)
+                end sub
+            `);
+
+            await process();
+
+            fsEqual(`${hostDir}/source/roku_modules/logger/main.brs`, `
+                sub logger_PrintValue(value)
+                    print bravo_printer_v1_writeLine(value)
+                end sub
+            `);
+        });
+
     });
 
 });
